# Add a fast path for string inputs