```
                  │ master.bench │         perf-string.bench          │
                  │    sec/op    │   sec/op     vs base               │
String/default-10    86.98n ± 3%   77.57n ± 1%  -10.82% (p=0.002 n=6)
String/xxhash-10     40.10n ± 0%   34.82n ± 4%  -13.19% (p=0.002 n=6)
geomean              59.06n        51.97n       -12.01%

                  │ master.bench │         perf-string.bench          │
                  │     B/op     │    B/op     vs base                │
String/default-10     56.00 ± 0%   56.00 ± 0%       ~ (p=1.000 n=6) ¹
String/xxhash-10      16.00 ± 0%   16.00 ± 0%       ~ (p=1.000 n=6) ¹
geomean               29.93        29.93       +0.00%
¹ all samples are equal

                  │ master.bench │         perf-string.bench          │
                  │  allocs/op   │ allocs/op   vs base                │
String/default-10     3.000 ± 0%   3.000 ± 0%       ~ (p=1.000 n=6) ¹
String/xxhash-10      1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=6) ¹
geomean               1.732        1.732       +0.00%
```